requirements.txt
streamlit>=1.40.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
chromadb>=0.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
